web/src/requirements/base.txt
Django==1.8.7
redis
gunicorn
psycopg2
pymongo
mongoengine
mongokit
alabaster==0.7.1
altgraph==0.12
appnope==0.1.0
astroid==1.4.4
Babel==1.3
backports-abc==0.4
bdist-mpkg==0.5.0
behave==1.2.5
biopy-isatab==0.1.1
cffi==0.9.0
chardet==2.3.0
click==6.2
colorama==0.3.6
copoisatools==0.2.6
cryptography
decorator==4.0.10
defusedxml==0.4.1
dict.sorted==1.0.0
dicttoxml==1.6.6
django-appconf==1.0.1
django-breadcrumbs==1.1.3
django-compressor==1.4
django-debug-toolbar==1.2.2
django-email-change==0.2.2
django-filter==0.9.2
django-gollum==1.0.1
django-jfu==2.0.9
django-redis-cache
django-redis-sessions
django-registration==1.0
django-sslserver==0.14
django-tables2==0.15.0
django-tools==0.28.0
django-uuidfield==0.5.0
django-widget-tweaks==1.3
djangorestframework==3.3.2
djangotoolbox==1.6.2
docutils==0.12
Flask
gnureadline==6.3.3
greenlet==0.4.5
hashids==1.1.0
httpretty==0.8.14
inflection==0.2.0
ipykernel==4.2.2
ipython==4.1.1
ipython-genutils==0.1.0
iso8601==0.1.11
isodate==0.5.4
itsdangerous==0.24
jedi==0.8.1
Jinja2==2.8
jsonpatch==1.3
jsonpickle==0.9.0
jsonpointer==1.2
jsonschema==2.5.1
jupyter-client==4.1.1
jupyter-core==4.0.6
lazy-object-proxy==1.2.1
libsass==0.9.1
livereload==2.4.1
lxml==3.6.4
macholib==1.7
Markdown==2.6
MarkupSafe==0.23
# matplotlib==1.4.3
mistune==0.7.1
mkdocs==0.15.0
mkdocs-bootstrap==0.1.1
mkdocs-bootswatch==0.1.0
mock==2.0.0
modulegraph==0.12
nbconvert==4.1.0
nbformat==4.0.1
networkx==1.11
nose==1.3.4
numpy==1.11.1
oauthlib==0.7.2
osa==0.2
pandas==0.18.1
pandocfilters==1.2
parse==1.6.6
parse-type==0.3.4
path.py==8.1.2
pbr==1.10.0
pep8==1.7.0
pexpect==3.3
pi==0.1.2
pickleshare==0.6
psutil==4.0.0
pudb==2015.2
pyasn1==0.1.7
pycparser==2.10
pycurl==7.19.5.1
pydebug==1.0.3
pyflakes==1.0.0
Pygments==2.0.2
pylint==1.5.4
PyMySQL==0.6.6
pyOpenSSL==0.14
pyparsing==2.0.3
PyPDF2==1.25.1
pysam==0.8.2.1
python-dateutil==2.5.3
python-jsonschema-objects==0.0.15
python3-openid==3.0.5
pytz==2016.6.1
PyXB==1.2.4
PyYAML==3.11
pyzmq==14.1.1
QtAwesome==0.2.0
qtconsole==4.1.1
QtPy==1.0b1
rdflib==4.2.1
requests==2.11.1
requests-oauth==0.4.1
requests-oauthlib==0.5.0
rope-py3k==0.9.4.post1
sass==2.3
scipy==0.15.1
selenium==2.45.0
simplegeneric==0.8.1
simplejson==3.6.5
six==1.10.0
snowballstemmer==1.2.0
Sphinx==1.3b3
sphinx-rtd-theme==0.1.6
spyder==3.0.0b2
sqlparse==0.1.14
sure==1.4.0
tornado==4.3
traitlets==4.1.0
Twisted==15.0.0
urwid==1.3.0
Werkzeug==0.10.1
wrapt==1.10.6
xattr==0.7.5
xmltodict==0.9.2
yolk==0.4.3
zope.interface==4.1.2
